Start of Firebase and some board changes
Added a few new things to the board creation page such as an ability to add a title. (i can soon make it so we can display text on the screen)

attempted to connect to firebase just for user authentication but it did not work. there is a button on the profile page that I added to test this, but again, broken.